test: 隔离全局环境与围栏宿主初始化
