+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- 测试环境初始化不再预加载生产模块；围栏宿主由需要它的测试显式加载，图表测试可直接使用普通模块 mock（#115）。
+
 ### 兼容性
 - 下一版 dsh-genui 的 DSH 支持范围为 `^0.1.2-rc.1 || ^0.1.3-alpha.1`。使用 DSH `<=0.1.1-rc.x` 的用户请固定使用 dsh-genui `0.9.8`。
