+++ tests/genui-gallery.spec.tsx
@@ -5,7 +5,7 @@
 // existing type, this file catches it.
 import { cleanup, render } from '@testing-library/react'
 import { afterEach, describe, expect, it } from 'vitest'
-import { hasFenceRegistry } from './setup'
+import { hasFenceRegistry } from './helpers/fence-host'
 import { MarkdownText } from './markdown-labels.tsx'
 import { gallerySpec } from '../src/client/gallery.ts'
 import { repairGenuiSpec } from '../src/client/guard.ts'
