+++ tests/setup.ts
@@ -1,25 +1,4 @@
-/**
- * Test setup: register the dsh-ui fence renderer exactly like the cordis
- * client entry does, so MarkdownText renders fences in jsdom — but ONLY when
- * the host under test ships the fence-registry extension point (contract
- * line). On pristine hosts the registry does not exist and the MarkdownText
- * fence integration tests skip themselves; the DOM channel has its own spec.
- * Also stubs requestAnimationFrame for the staggered-reveal animation and
- * SVG measure APIs jsdom lacks.
- */
-import * as primitives from '@deepseek-ai/dsh-client-ui-primitives'
-import { renderGenuiFence } from '../src/client/index.tsx'
-
-type HostFenceExt = {
-  registerFenceRenderer?: (lang: string, renderer: (raw: string, key: unknown, context?: unknown) => unknown) => () => void
-}
-
-export const hasFenceRegistry: boolean = typeof (primitives as unknown as HostFenceExt).registerFenceRenderer === 'function'
-
-if (hasFenceRegistry) {
-  ;(primitives as unknown as HostFenceExt).registerFenceRenderer!('dsh-ui', renderGenuiFence as never)
-}
-
+/** Browser API polyfills for the test environment. */
 // jsdom lacks rAF: the reveal animation uses it per item; manual tick below.
 if (typeof globalThis.requestAnimationFrame !== 'function') {
   // @ts-expect-error test-only stub
